Separate IP name lookup implementation out from the WASI translation layer
